+++ src/d_deh.c
@@ -217,6 +217,8 @@ char    *s_PD_YELLOWS = "";
 char    *s_PD_ANY = "";
 char    *s_PD_ALL3 = "";
 char    *s_PD_ALL6 = "";
+char    *s_KEYCARD = "";
+char    *s_SKULLKEY = "";
 
 char    *s_SECRET = "";
 
@@ -784,6 +786,8 @@ deh_strs deh_strlookup[] =
     { &s_PD_ANY,                     "PD_ANY"                     },
     { &s_PD_ALL3,                    "PD_ALL3"                    },
     { &s_PD_ALL6,                    "PD_ALL6"                    },
+    { &s_KEYCARD,                    "KEYCARD"                    },
+    { &s_SKULLKEY,                   "SKULLKEY"                   },
 
     { &s_SECRET,                     "SECRET"                     },