+++ src/p_doors.c
@@ -262,7 +262,7 @@ dboolean EV_DoLockedDoor(line_t *line, vldoor_e type, mobj_t *thing, fixed_t spe
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_BLUEO, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "keycard");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_KEYCARD);
                     HU_PlayerMessage(buffer, false, false);
                 }
                 else if (player->cards[it_blueskull] == CARDNOTFOUNDYET)
@@ -274,7 +274,7 @@ dboolean EV_DoLockedDoor(line_t *line, vldoor_e type, mobj_t *thing, fixed_t spe
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_BLUEO, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "skull key");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_SKULLKEY);
                     HU_PlayerMessage(buffer, false, false);
                 }
 
@@ -299,7 +299,7 @@ dboolean EV_DoLockedDoor(line_t *line, vldoor_e type, mobj_t *thing, fixed_t spe
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_REDO, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "keycard");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_KEYCARD);
                     HU_PlayerMessage(buffer, false, false);
                 }
                 else if (player->cards[it_redskull] == CARDNOTFOUNDYET)
@@ -311,7 +311,7 @@ dboolean EV_DoLockedDoor(line_t *line, vldoor_e type, mobj_t *thing, fixed_t spe
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_REDO, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "skull key");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_SKULLKEY);
                     HU_PlayerMessage(buffer, false, false);
                 }
 
@@ -336,7 +336,7 @@ dboolean EV_DoLockedDoor(line_t *line, vldoor_e type, mobj_t *thing, fixed_t spe
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_YELLOWO, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "keycard");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_KEYCARD);
                     HU_PlayerMessage(buffer, false, false);
                 }
                 else if (player->cards[it_yellowskull] == CARDNOTFOUNDYET)
@@ -348,7 +348,7 @@ dboolean EV_DoLockedDoor(line_t *line, vldoor_e type, mobj_t *thing, fixed_t spe
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_YELLOWO, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "skull key");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_SKULLKEY);
                     HU_PlayerMessage(buffer, false, false);
                 }
 
@@ -511,7 +511,7 @@ void EV_VerticalDoor(line_t *line, mobj_t *thing)
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_BLUEK, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "keycard");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_KEYCARD);
                     HU_PlayerMessage(buffer, false, false);
                 }
                 else if (player->cards[it_blueskull] == CARDNOTFOUNDYET)
@@ -523,7 +523,7 @@ void EV_VerticalDoor(line_t *line, mobj_t *thing)
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_BLUEK, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "skull key");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_SKULLKEY);
                     HU_PlayerMessage(buffer, false, false);
                 }
 
@@ -551,7 +551,7 @@ void EV_VerticalDoor(line_t *line, mobj_t *thing)
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_YELLOWK, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "keycard");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_KEYCARD);
                     HU_PlayerMessage(buffer, false, false);
                 }
                 else if (player->cards[it_yellowskull] == CARDNOTFOUNDYET)
@@ -563,7 +563,7 @@ void EV_VerticalDoor(line_t *line, mobj_t *thing)
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_YELLOWK, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "skull key");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_SKULLKEY);
                     HU_PlayerMessage(buffer, false, false);
                 }
 
@@ -591,7 +591,7 @@ void EV_VerticalDoor(line_t *line, mobj_t *thing)
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_REDK, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "keycard");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_KEYCARD);
                     HU_PlayerMessage(buffer, false, false);
                 }
                 else if (player->cards[it_redskull] == CARDNOTFOUNDYET)
@@ -603,7 +603,7 @@ void EV_VerticalDoor(line_t *line, mobj_t *thing)
                     }
 
                     M_snprintf(buffer, sizeof(buffer), s_PD_REDK, playername,
-                        (M_StringCompare(playername, playername_default) ? "" : "s"), "skull key");
+                        (M_StringCompare(playername, playername_default) ? "" : "s"), s_SKULLKEY);
                     HU_PlayerMessage(buffer, false, false);
                 }