+++ releasenotes.md
@@ -3,6 +3,7 @@
 * Optimizations have been made to further improve the overall performance and stability of *DOOM Retro*.
 * Minor changes have been made to the character set used in the console.
 * Minor changes have been made to text that is output to the console.
+* A warning will now be displayed in the console if a locked door is found with no keycard or skull key in the map to open it.
 
 ---